Fixed wrong relocation ://
