+++ Essentials/src/main/java/net/ess3/api/events/teleport/TeleportEvent.java
@@ -20,7 +20,7 @@ public abstract class TeleportEvent extends Event implements Cancellable {
     private final ITarget target;
     private boolean cancelled = false;
 
-    protected TeleportEvent(final IUser teleporter, final IUser teleportee, final PlayerTeleportEvent.TeleportCause cause, final ITarget target) {
+    TeleportEvent(final IUser teleporter, final IUser teleportee, final PlayerTeleportEvent.TeleportCause cause, final ITarget target) {
         super(!Bukkit.isPrimaryThread());
         this.teleporter = teleporter;
         this.teleportee = teleportee;
@@ -28,7 +28,7 @@ public abstract class TeleportEvent extends Event implements Cancellable {
         this.target = target;
     }
 
-    protected TeleportEvent(final IUser teleportee, final PlayerTeleportEvent.TeleportCause cause, final ITarget target) {
+    TeleportEvent(final IUser teleportee, final PlayerTeleportEvent.TeleportCause cause, final ITarget target) {
         this(teleportee, teleportee, cause, target);
     }
