+++ Essentials/src/main/java/net/ess3/api/events/teleport/TeleportWarmupEvent.java
@@ -1,8 +1,7 @@
-package net.essentialsx.api.v2.events;
+package net.ess3.api.events.teleport;
 
 import com.earth2me.essentials.ITarget;
 import net.ess3.api.IUser;
-import net.ess3.api.events.teleport.TeleportEvent;
 import org.bukkit.event.HandlerList;
 import org.bukkit.event.player.PlayerTeleportEvent;
